Fine Tuned Monoliths Center
Adjusted Home Position and Probe offset with a peice of scotch tape and an X marking center. Probing and a G1 X117.5 Y117.5 both land On the center of the X Post tuning.

May be useful on a permachine basis curious how many Enders are off my 3mm on the X as the Hero Me did not change x positions and I did use to compensate when slicing,